Depend on kernel UAPI instead of libbtrfs
Based on moby/moby PR 44761. (Thanks to [@]neersighted.)

The containerd/btrfs library now requires headers from kernel 4.12 or newer:
- https://github.com/torvalds/linux/blob/master/include/uapi/linux/btrfs.h
- https://github.com/torvalds/linux/blob/master/include/uapi/linux/btrfs_tree.h

These files are licensed under the GPL-2.0 WITH Linux-syscall-note, so
it should be compatible with the Apache License 2.0.
https://spdx.org/licenses/Linux-syscall-note.html

The dependency on the kernel headers only affects users building from source.
Users on older kernels may opt to not compile this library, or to provide headers
from a newer kernel.

To compile containerd/containerd without btrfs, grep `no_btrfs` in
https://github.com/containerd/containerd/blob/main/BUILDING.md

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>